+++ packages/svelte2tsx/test/svelte2tsx/samples/runes-bindable-reserved-name.v5/expectedv2.ts
@@ -1,0 +1,10 @@
+///<reference types="svelte" />
+;function $$render() {
+
+    let/** @typedef {{ a: any, class?: any }} $$ComponentProps *//** @type {$$ComponentProps} */ { a, class: className = $bindable() } = $props()/*Ωignore_startΩ*/;className;/*Ωignore_endΩ*/;
+;
+async () => {};
+return { props: /** @type {$$ComponentProps} */({}), exports: {}, bindings: __sveltets_$$bindings('class'), slots: {}, events: {} }}
+const Input__SvelteComponent_ = __sveltets_2_fn_component($$render());
+/*Ωignore_startΩ*/type Input__SvelteComponent_ = ReturnType<typeof Input__SvelteComponent_>;
+/*Ωignore_endΩ*/export default Input__SvelteComponent_;