+++ packages/svelte-check/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 4.4.8
+
+### Patch Changes
+
+-   feat: typescript 6.0 support ([#2985](https://github.com/sveltejs/language-tools/pull/2985))
+
+## 4.4.7
+
+### Patch Changes
+
+-   fix: flush stdout/stderr before exit ([#3014](https://github.com/sveltejs/language-tools/pull/3014))
+
+-   fix: report diagnostics in tsconfig.json ([#3005](https://github.com/sveltejs/language-tools/pull/3005))
+
 ## 4.4.6
 
 ### Patch Changes